:sparkles: [Feat] 공통함수 분리
