+++ src/components/card/UserProfile.module.css
@@ -80,7 +80,7 @@
 .driverName {
   font-size: 24px;
   font-weight: 600;
-  color: var(--balck-300);
+  color: var(--black-300);
 }
 
 .movingInfoRType {
